ethernet:lpc32xx: Ethernet PHY timeout fix
The lpc_mdio_read() function returns an EIO before the final
lpc_net_hard_start_xmit() is called. It turns out that this is because
jiffies rolls over 5 minutes after boot. On a build with my gcc-4.6.1
toolchain, the phy just happens to get polled at a point when jiffies is
0xFFFFFFF8. Timeout therefore becomes +2 and the timeout comparison
fails immediately. The function returns an error and does not set
LPC_ENET_MCMD back to 0. I guess this means that the phy can no longer
receive packets, hence the crc errors. The atached patch fixes the
issue.

Signed-off-by: Bangaragiri G <bangaragiri.g@nxp.com>